src/requirements.txt
cloudshell-shell-core>=3.1,<3.2
cloudshell-automation-api>=7.1,<7.2
cloudshell-tg-breaking-point>=1.0,<1.1
cloudshell-networking-devices>=1.0,<1.1
nahum-debug-utils
requests
retrying